Rework Studio Help integration
Commit 73f738aad9d fixed the registration for StudioHelpManagerImpl
which was taking over IntelliJ's help from within the Android plugin.
Now as a Studio-only service defined in adt-branding, it is loaded by
the platform classloader, but the code is not accessible to the
platform classloader.

This commit takes the opportunity to simplify the implementation and
avoid the problems above by implementing a new WebHelpProvider
extension, which handles the Android specific help links (such that
they work outside Studio too) and leaves the rest to the platform.

Bug: 158445825
Bug: 153337903
Test: AndroidStudioHelpTest
Change-Id: I66f64266bca246fa772f4f52eedfeb8faf8acd20

GitOrigin-RevId: a6b520d16e858758579debb491b17041694defb2