Fix Logcat disposal during project closing
When 2 or more projects are open and Logcat is active (for instance
there's a running emulator), attempting to close one project leaves
the IDE with a partially-disposed, empty project window which can't
be closed anymore, and the exception looks like:

Cannot create ServiceAdapter(
  descriptor=ServiceDescriptor(
   interface=com.intellij.openapi.util.WindowStateService,
   implementation=com.intellij.openapi.util.WindowStateProjectService
  ),
  pluginDescriptor=PluginDescriptor(
    name=IDEA CORE,
    classpath=C:\Program Files\Android\Android Studio\lib\resources.jar
  ))
because container is already disposed (container=Project (
  name=...,
  containerState=DISPOSE_IN_PROGRESS,
  componentStore=...)
) [Plugin: com.intellij]

This happens because we attempt to dispose Logcat too late, when other
project services have already been disposed, during ProjectManagerImpl.
fireProjectClosed(). And unfortunately the thrown exception prevents
the project closing.

Similar components using a RunnerLayoutUI include the build and run
tool windows; like BuildContentManagerImpl or RunContentManagerImpl,
we should provide a closeQuery and participate in early disposal, so
that BaseContentCloseListener.canClose() performs the disposal for us
during the ProjectManagerImpl.fireProjectClosing() call.

Bug: 157825209
Test: manual
Change-Id: Iecea245b25e5b872bac9e5e9db2f3bcb7189327c

GitOrigin-RevId: 564a6de3dc675cc82b233ff99af8b4dac9bb2186