[BA] Fix background color update problem on theme change
Color attributes were not updated properly on color theme change
because page views were created but not added the the components tree
right away. Because of that some panels were created with the initial
theme parameters before the change but could not receive the new theme
parameters when applying the theme change because theme update code
was not aware of such components.
This happened because views on the main panel were created eagerly on
main view init but CardLayoutPanel adds them to the components tree only
on the first 'select'.
Fix this problem with selecting all three pages on the main view init
to trigger all pages addition to the components tree.

Bug: 157052075
Test: visually verified
Change-Id: Iecccb347ea99fbdf41c60db6951e0b0be54aa4e5

GitOrigin-RevId: bcca907fbd5dd527dfab095c9e38c59cd09dd49b