Refactor AppInspection API
The main purpose of this CL is to reorganize and refactor AppInspectionDiscovery. AppInspectionDiscovery grew very large and became a do-everything class. This CL spilts it up into 3 classes:
1) ProcessNotifier - susbscription class used by UI to listen for new processes
2) AppInspectorLauncher - a simple entry point for launching an inspector on a process
4) AppInspectorTargetManager - internal class used by AppInspectorClientManager to attach to processes

Only 1 and 2 are exposed to the public via AppInspectionApiServices interface. In addition, AppInspectionTarget is now responsible for keeping track of all clients launched on that target.

This CL also cleans up the tests by putting more setup code in AppInspectionServiceRule.

Test: Moved AppInspectionDiscovery tests to new test classes. Added new test for AppInspectorLauncher.
Bug: 158933346
Change-Id: I61b3d9c0eef8846fb8191bde6419335fed498b00

GitOrigin-RevId: 4e102706adc48583e52ef1491eb0e9c0d38737ef