[Build Analyzer] Disable non incremental APs warnings when using kapt
When users declare annotationProcessor dependency on an annotation
processor instead of/with using kapt configuration while applying kapt,
JavaCompile task receives static information about the AP
incrementality and doesn't end up running the APs. This means in case
of a dynamic AP such as room or databinding, JavaCompile task sends to
the tooling API that they're not incremental although they might have
been.

As a workaround this change disabled warnings about non incap in the
build analyzer when running kapt.

Bug: 157673275
Test: updated AnnotationProcessorsAnalyzerTest
Change-Id: I59d410513dd2d3702085f86392f1b3208a18d510

GitOrigin-RevId: 88eec3000e1c7cc835fc60c50a14f7b346d5ea23