Convert the animation inspector to a tabbed pane
Instead of a single panel, the animation inspector should have a tabbed
pane, and each animation needs to be represented by its own tab. The tab is
created when the animation is subscribed and removed when it's unsubscribed.

Each tab has its own toolbar, states combobox and animated properties panel.
All the tabs share the same timeline, so users can see how animations
interact with each other.

If there are no animations subscribed, we display a panel telling it so.

Bug: 157895086
Test: UI test TBD when the inspector UI is more stable
Change-Id: Iceed0feb73ae9339b521da32fc8a16c8050e2067

GitOrigin-RevId: 92262467855e01f680f9c2f624860d32dbab7c75