Delete AndroidXmlLanguageCodeStyleSettingsProvider
It does not appear to be necessary, because we already apply
AndroidXmlPredefinedCodeStyle from within the CodeStyleSettings
constructor in the platform (see Change I29434c6c2).

Plus, AndroidXmlLanguageCodeStyleSettingsProvider caused bugs
because the platform does not expect there to be multiple
LanguageCodeStyleSettingsProviders registered for the same language.
For example, one bug is that there were duplicate XML sections listed
under Settings -> Editor -> Code Style.

See go/studio-code-style-settings for more background and notes.

Fixes: 157253341
Test: existing (AndroidCodeStyleSettingsTest), and manual
Change-Id: I0bdb9f623a9a8cae0b7b6e9ed52eea8651a104e1

GitOrigin-RevId: 455c8d8bae400d003bbe439d05c82b45bb94b2fa