Better fix that doesn't swallow all errors
This CL amends an earlier fix that ignored all exceptions.
Instead of ignoring all exceptions, we only ignore
ProcessCanceledExceptions, and re-throw the first non-PCE
if there is one. This is to ensure all post-invocation tasks
still run, but what normally would be the first non-PCE
isn't lost to the wind.

Fixes: 158784803
Test: N/A
Change-Id: Idaf7c055ceaffc7ba78f9acde928a1202de1293e

GitOrigin-RevId: 2b5a960e39e87a3d85dfda2345cf5af228acda9f