Remove unsafe parent disposer
AndroidJavaDebugger registers one of its internal listeners as a
disposer tree dependency under the current Project. This is
unnecssary and causes a false dependency on the Project, preventing
the listener from being disposed properly. The listener will will be
removed once the RunContentDescriptor is disposed via a lifetime
strictly less than the Project, as the RunContenDestriptor
is the one holding onto the ProcessHandler, which holds onto the
listener.

Fixes: 159168468
Test: N/A
Change-Id: Ic3fab82323fa98923796cb05476d706e032882bb