[Dsl] implement rename() for configurations
In this CL, rename() acts like flavor renames (and unlike
signingConfig renames) in that there is no option to attempt to track
uses of the configuration, but performs purely a syntactic rename on
the element that introduces the configuration.  A future CL is likely
to add the option of also renaming references (to the best of the
ability of the Dsl parser to achieve this).

Bug: 149565531
Test: added
Change-Id: I148696e915eab080ce766b4fb2dee218d27477e1

GitOrigin-RevId: 863da99aafe21aed34b5319f935ae1939645a501