Avoid capturing surface in lambda
To prevent a memory leak, avoid capturing the surface as part of the
lambda given to the CommonUsageTrackerImpl. We only need the
applicationId so we do not need to depend on the surface once we have
calculated it.

Fixes: 159159201
Test: N/A
Change-Id: Id0afb1b1faf5a921dc6c7570cbea2a145b95df3d
(cherry picked from commit 7be3abae3244c74066ad64fc5dbb26d0459374da)