Fix the zoom-to-fit and zoom-to-actual actions in DesignSurface
Fixed DesignSurface can still zoom-to-fit or zoom-to-actual on some
displays even the zoom level is already matched.

Change the type of screen scaling factor from float to double to match
the type of scaling value.

Also add some annotations for the scaling values in DesignSurface to
improve the readability of relationship between scaling value and zoom
level.

Fixes: 158041314
Test: manually / covered by current tests.
Change-Id: Id91f59eb972f696428453ee3d550ae7dd833be25

GitOrigin-RevId: 40492dc44777c5481da13bc5b272c76c2026a5a1