Click on a test result row opens a details page
TreeTableView does not call valueChanged callback when you
click on a row which has been selected already. This is
undesirable because you may want to reopen the details page
by clicking on the same row. This patch calls the registered
callback from the mouse listener in addition to the valueChanged
callback.

Bug: n/a
Test: AndroidTestResultsTableViewTest
Change-Id: I7ef44cdf22cd2287ff2f16f1d5058fb09a6c7399