[safeArg] Provide class-based package descriptors
Since descriptors generated and stored in
KtDescritporCacheModuleService don't necessarily
belong to the module where coresponding nav resource
files sit, extra work are needed to associate descriptors
for each of the module.

Also, since package names of generated classes are
not the same with the module name(it can be from
declared fragment class), so we can't provide one
package descripotor per module, instead,
we need to provide a bunch of package descriptors,
and it contains one args/directions class for each
of the package descriptor.

Performance should be monitored in the future, though
ProgressManager#checkCanceled are invoked as frequently
as possible.

Bug: 158507086
Test: Existing tests updated
Change-Id: I3ea92cdb35acf832fcec14f725e7025f81a9cf3a

GitOrigin-RevId: a82e69fc9240c711f431b349d65fe30edc324638