Fix flickering due to setting layout_marginLeft when not needed.
When MinApi is 17 or above we shouldn't set layout_marginLeft and
layout_marginRight. It generates a lint warning and the declared
attributes filters it out.

This CL makes the ConstraintWidget only set the rtl variant of the
margin attributes when the MinAPI is 17 or above. It continues to
set both when the MinAPI is 16 or below.

Fixes: 159474602
Test: Unit tests added.
Change-Id: Ic3c88efa3c4df8261296f1c71e295c59adf3f893

GitOrigin-RevId: 047b40e771572cfe7377dc0d6617fefd1c1d9bf0