[safeArg] Provide completion contributor for kt
Completions provided here are targeting those class
descriptors from not direct directives in module
with all dependencies.

For completions in imported directive expressions
or with reciver expressions, it relies on original
KotlinCompletionContributor.

Ideally, all duplications should be removed. But in
this cl, we only remove two easy cases, 1) if the package
name of containing file is exactly matching with
package name of the class descriptor result. 2) if
the package of the class descriptors result is already
imported. Maybe some other duplications would be taken
care in the future cls.

Bug: 159808741
Test: SafeArgsKtCompletionContributorTest
Change-Id: Ieebf3d51641641b069a926e131905d7b4a3812fe

GitOrigin-RevId: 6e29ed110cb2428accbc055cdce3b84351eba5fb