[DSL] Utility function to build a GradlePropertyModel.
The added functions allows creating and building a GradlePropertyModel
given the name of the  dsl property we want to fetch.
getModelFromExternalText() is marked as deprecated because this will
only be used for the temporary solution for ReferenceTo until we
completely delete setting references from String values.
getModelFromInternalText() deals with internal names (DSL names) for
lookups.

Bug: 14254204
Test: added
Change-Id: Ib7d468cf65786ccadfd46e25aa73e05e288b1143

GitOrigin-RevId: 0e48d54e6eb1228333aa12493db2ec449eacfdec