[DSL] new ReferenceTo API.
Introduce a new way of setting references given a property context. This will only allow having resolvable DSL references. The property context is needed to resolve references accurately.
For example, if we have a property name concurrency in a project, we need to make sure the reference is sat to the property with the closest DSL scope, and this isn't guaranteed with a bottom-up / top-down approach without a context.

Bug: 142454204
Test: N/A
Change-Id: I229dc2a46277a39ef2954e863b2bae9c95e90b31

GitOrigin-RevId: 527583f25f4f07540eddb0674a7b2328d31ad902