Fix animation inspector timeline offset and trailing effect
The timeline marker was being drawn slightly to the left of the actual
value. We fixed that by taking into account the thumbRect and drawing the
thumb in the middle of it, not aligned to the left.

Also, there was a trailing effect when scrubbing the timeline, resulting on
a horrible UI when dragging the timeline marker around. That was fixed by
creating our own TrackListener that updates the thumbLocation on mouse
pressed and dragged.

Bug: 157895086
Test: Visual Change. N/A.
Change-Id: I06adbda3c4b1819955eb7c1b5a6e9eca5a5eb1d0

GitOrigin-RevId: 714858d7b34bdc8b3b9c9026afb0736f55584087