Add scheduled device to test matrix before test execution
Passes IDevice to AndroidTestApplicationLaunchTask constructor
and create a DdmlibTestRunListenerAdapter. The adapter updates
the test matrix. Before this patch, the adapter is created in
AndroidTestApplicationLaunchTask.run method which is called
after the APK installation task. Because the installation task
is very expensive, the device name is not displayed to users
until the very last moment of test execution. This is especially
problematic when you run tests on two devices where one is a new
device and the other one is old.

Bug: n/a
Test: manual and unit tests
Change-Id: I6255411813723aa169104b509151bb13e34f46f9

GitOrigin-RevId: 94ddb9bad19580f91f29d3ec82636ee35d3960fb