[AUA] refine computation of applicability
Following discussions, identify and compute for each refactoring
processor not the applicability of the transformation, but a more
finely-graduated reflection of its necessity for a given AGP version
upgrade: is a particular transformation (manifest in a refactoring
processor) mandatory, optional or irrelevant, and is it independent of
the upgrade (i.e. can be done before) or is it dependent (must be done
at or possibly after)?

Initialize isEnabled as true for all the non-irrelevant refactoring
processors for now, though this may change to include only the
mandatory processors in the near future.  Likewise, in the current
version of the AgpUpgradeRefactoringProcessorDialog, set the initial
selection from all non-irrelevant processors; again, this may change
soon.

Bug: 149565531
Test: existing (no change in functionality)
Change-Id: Ic6ed923314d5a6313d685943398b8010b4a9e243

GitOrigin-RevId: 115e1aba37195fbb634f621305955f6e3397b03d