[AUA] dialog for Canary 5
Construct a dialog box for "MVP" graphical AGP Upgrade Assistant,
which for the 4.2 cycle, or at least 4.2.0-alpha05, will have only one
refactoring with settings, namely the processor which handles the
change of default Java from 7 to 8.  Construct a dialog which only
offers users the choice of accepting the new default or inserting
directives to preserve the old behaviour.

Bug: 159612385
Test: manual
Change-Id: I2b45f2025bba5bd5f52428d53d5187e648e9f526

GitOrigin-RevId: 54a6c82f1982ba17716cce983790a3822d774f8a