[AUA] a setting for the Java8Default refactoring
Specify the action the processor should take if there is no explicit
language level statement, one of: inserting directives corresponding
to the old default, so that the existing behaviour is preserved, or
leaving the build files alone, accepting the new default behaviour.

Since this now provides multiple possibilities for the refactoring
action's effect, modify the associated command name and the processed
elements header accordingly.

Bug: 159612385
Test: added
Change-Id: I089d305795abb1dfff8ed5cfccaa3a6d842d14ea

GitOrigin-RevId: 71da134fd3799f3c2d34e37a82eddc76be71f7e2