Fix DeviceConfiguratorPanel layout
There are two main panels. On the left
"Available qualifiers"/"Chosen qualifiers",
and on the right the "editor" for the
"Chosen qualifiers".
The left panel was calculating a preferred size,
but because was added to the CENTER of a BorderLayout
panel, it was forced, by the right panel (BorderLayout.EAST),
to take any remaining space. If the right panel width is
"large" (eg "Locale editor"), then it squashes the left
panel.
This CL sets the width weight of the right panel to 50%.

Bug: 37105561
Test: Manual inspection of the UI layout
Change-Id: I08cf04dce5c00c12f82d33083ceae77a22a64385

GitOrigin-RevId: 0ff7f3aecd34ab1fc215b200ec1ef3fc75e5ba12