Stop building skia-grpc-server by default
In general there's no reason this needs to be part of the studio build,
since in the normal case the version from the sdk is used.
Also, once skia is updated, the "studio" target won't be able to build
it until a lot of updates are done one the machines that do that build.

Test: none
Bug: none
Change-Id: I9edabd5ae0f6bd886a61bfa9d54af31e972512b3

GitOrigin-RevId: 11393d37a71ca52100742376273fdea770c1f690