Simplify tests using a custom Kotlin class visitor
Instead of reflection, Kotlin descriptors accept visitors, and the
Kotlin plugin supplies a handful of its own renderers, which visit
descriptors and generate rich text as a result.

However, because we're in a test environment, some classes we reference
don't resolve. But more than that, the renderers seem somewhat
arbitrary, and there's no guarantee that they won't be tweaked in the
future, so it seems a bit risky to rely on their text output as a way
to verify class fields.

So for this CL, we create our own custom visitor, which should guarantee
a bit of stability, as well as make the tests slightly more readable,
as we can reduce boilerplate code across tests.

Bug: N/A
Test: Only test code updated
Change-Id: Id63ca2c1377f6fca45104ebfaca374a5e13624e6

GitOrigin-RevId: 43829442b536be25e21966ea5b1fb4c43cac3025