Set to null null cells edited with empty string
When a cell contains the NULL value and is edited we currently
commit the empty string, overriding the NULL value.
This is wrong when the user just double clicks on a cell
and clicks away without intending to edit the cell.

This CL changes this behaviour so that if a cell
containing the NULL value is edited with the empty
string, the null value is maintained.

The drawback of this solution is that it's not possible
to just edit a cell and replace NULL with empty string.
It has to be done by running an UPDATE query or editing
the cell to something different from the empty string first.
This limitation however is common in other tools and doesn't
seem to be a big issue for users.

Test: TableViewImplTest.kt
Bug: 159993801
Change-Id: I652aedd0584da7e1577181d9321f8c9a48aac3b8
(cherry picked from commit 487235079d6e99b9816b5b0f8dc5e3f3387dc50f)