Update skia parser to latest version
We need to make sure the skia parser is update to a compatible version.
This will make the compatibility story much simpler: instead
of having to worry about handling different versions on both
the parser side and the studio side, we can assume on the
studio side that we have the latest revision of the skia parser,
per package (that is, skiaparser;1 and skiaparser;2 will be
different, but if we want we can update skiaparser;1 itself).

Bug: 150195151
Test: none
Change-Id: Ia02cb4986c5f64f776c960500e1f94baf014fa31
(cherry picked from commit c71e7f4261599a0562cb11b096c19506305989ed)