[Build Analyzer] Change AGP report file location to projectDir/.gradle/
Having a temp folder for the report that gets published by AGP that
changes per build means the injected file location property changes
per build and so Gradle can't reuse the config cache when configuration
caching is enabled. This changes the report file location to be in
a specific directory instead.

Bug: 158197785
Test: existing build analyzer integration tests
Change-Id: Id1954dd9a72a97d8549e092c0059de23c304e9ca

GitOrigin-RevId: 388013784eb5d097bfe5ae0e2241dc910340acf1