Add support for basic end-2-end QrCode pairing
Pairing multiple device over Wi-Fi is supported end-2-end with this
change. When the pairing dialog shows up, it displays a QR Code that
can be scanned by the camera of a physical device to pair the device
over Wi-Fi. Once paired, the pairing dialog also waits for the device
to show up as a connected device to ensure the "pairing" was successful.

This change also includes a new AdbFileProvider class, which serves
as a general purpose way of accessing ADB from modules that don't
depend on android-core (they only need to depend on android-adb).

* This change does not include error handling related to checking
  platform requirements (adb initialization, platform support for
  mDNS, etc.)
* This change does not include pairing via PIN code
* This change does not include the final UI
* This change does not include unit tests. Unit tests will be added when
  the UI is closer to a finalized state.

Bug: 150796676
Test: Later
Change-Id: Ic95c6c77032fab6bf2595ab0064e587fa5af30c3

GitOrigin-RevId: ee3be7afb1084e467dc10f1bef7f16c6750b5c43