[BA] Fix viz scrollbar being rendered over the content
Problem:
Normally scroll bars are rendered over the content and hide after scrolling
stops. However in some setups they start being rendered next to the viewport
and occupy actual space making the viewport shrink.
But this implementation of visualization has fixed size an can not shrink
making part of it out of view.

Fix: making vertical scroll always on reserves extra space for it so that
it is no more rendered over the content.
Also set it forcefully to be non-opaque, otherwise empty scrollbar space
is visible (having a different background) even if not required.
Reduce margin between the visualization and details panels so that it remain
the same space with normal scrollbar width.

Just setting opaque=false would solve the issue as well as it being true
in some setups is the root cause of the issue. But as I am not sure
it is applied properly in all cases I still want to reserve dedicated space
for the scroll.

This is a temporal solution, visualization is moving next to the tree
in the next version.

Bug: 160170859
Test: Visually verified
Change-Id: I179a348f195c73f1101252ce8987c13529d6fb15
(cherry picked from commit a5d7f6c279b7619e0c4e031626fdbdf921b78102)