Fix Debug/Profile when multiple devices are selected
Due to an API change, launchable devices are now not return in
AndroidExecutionTarget#getDevices. This causes the check for multiple
devices to fail, and erroroneously enabling the Debug/Profile buttons
when multiple unlaunched emulators are selected.

The fix is to add an extra AndroidExecutionTarget API to return the
number of launchable (i.e. pre-execution) devices that are currently
selected.

Test: AndroidProgramRunnerTest
Fixes: N/A
Change-Id: I3441cd211a0476f63284fd473d09f9fadca4633c