Ensure databinding kapt errors are visible to the user
While using kapt.use.worker.api errors from databinding are not logged
to the Gradle output as is expected by the DataBindingOutputParser. This
change adds a new GradleIssueChecker that allows these errors to be
found by the user without re-running with --stacktrace.

There are a few drawbacks to this, firstly the InvokationTargetException
is still displayed as a useless item on the build output tree. Secondly
BuildIssueCheckers only allow us to create a single issue off the
exception, for databinding there can be multiple errors encoded. To get
around this we create a single build issue which is a concatanation of
all of the databinding errors that were found.

Bug: 156837289
Test: unit tests added
Change-Id: Ida4f02798085edf455ad52b3cf75e6a0fa50378a