Start deleting reference settup using String vals.
Remove the ability to set references to a text value. With this change we can now only set DSL resolvable references. For any unresolvable reference text, RawText is going to be the better solution for that. as the text (myReferenceText) refers to the fullyQualifiedName of the referred property, allow getting the simple property name through toString(). This is what clients should use of the reference.

Bug: 142454204
Test: updated.
Change-Id: I91077a591c03cee99308fab215113f78ce92dbe1