Reset tabs count when removing all animation inspector tabs
When removing all the animation inspector tabs while it's open, which can
happen if all the animations are unsubscribed or (most likely) when a build
or refresh is triggered, we should reset the tabs counter to prevent it from
growing too much.

The tab counter basically determines the tabs title, which are formatted as
"AnimationType #Counter", e.g. "TransitionAnimation #1". Resetting the
counter when there are no more tabs open provides a better UX for user,
since it's easier to keep track of lower number.

For instance, if the user starts inspecting 5 animations, there are 5 tabs
named #1 to #5. Every rebuild while the inspector is open would replace
these tabs with new tabs named #6 to #10, #11 to #15 and so on. Resetting
the counter will make the tabs start to be named from #1 whenever it's
possible.

Bug: 157895086
Test: N/A. Simple visual change that only affects the tab titles.
Change-Id: I2581f53061cc8d06a9ced9b2b1f0938d540a34f6

GitOrigin-RevId: 14ee27790783808b4fbc5a2384ec64721f9c640d