[AUA] is it certain that a refactoring is a no-op
It is possible to tell ahead-of-time whether a given refactoring can
ever cause a change in the project.  By default, we assume that:

- perform() / doRefactoring() is a function only of the usages;
- the usages found are independent of any settings in the refactoring.

Under those conditions, the refactoring is guaranteed to be a no-op if
no usages are found.  Define a protected method
computeIsAlwaysNoOpForProject() for AUA component processors with a
default implementation reflecting this; use it to initialize a public
var isAlwaysNoOpForProject (whose setter is only for testing
purposes).

Some refactorings might find usages and yet nevertheless be guaranteed
not to change them; for example, the Java8Default refactoring returns
usages of existing language level properties, in order to show them in
the preview window.  Override computeIsAlwaysNoOpForProject() in that
class to handle this difference.

Use this function in the current AUA dialog box to hide the settings
drop-down (but not the existence of the processor itself) for the
Java8Default processor.

Fixes: 161077357
Test: added unit tests for processor and dialog box
Change-Id: Ie1954fa0606164a36ef839b079d5721d16d77736

GitOrigin-RevId: 1f51bd0978e75375169bb52fdd4eafb7ce72c19c