Differentiate between name and accelerator text
Apply (Code) Changes action "name" was being used for both its
textual name as well as the accelerator name (IJ markup to a
string to support accelerator keys). This was causing an issue
where the text description contained the markup as well.

This CL basically creates a new variable to hold the accelerator
name instead of piggybacking on the textual name.

Fixes: 158821567
Test: N/A
Change-Id: I20bd9197296153762af2cb7be655313f8dc6d73b