Fix animation inspector labels painting when changing state
It is possible to define transitions between particular states in an
animation. Therefore, the animation inspector timeline might have different
maximum values when an animation is going from state A to state B and when
it's going from A to C. We were not handling this scenario appropriately,
and the labels showing the time were not being repainted with the correct
time.

Test: N/A. Visual change.
Bug: 157895086
Change-Id: I5d517bec3bc2c047bba6e2d025be484e4a5a5995