[safeArg] Support overridden args in an action
Both overriding arguments and arguments from destination
are collected. We fallback to NavDirections instead of
creating new inner action builder class if parameter
list is empty. Please note that parameters with default
value being declared are filtered out.

If incompatible types of an argument occur, we surface
issues with warnings, but still provide the best results,
and let it fail when compiling.

This CL only fixes issues for java light classes. Fixes
for kt descriptors would be in a follow up CL.

Bug: 161353845
Test: Updated existing test
Change-Id: Ia350317e7a17ee9cc295de7a8a46cdbce032bde1

GitOrigin-RevId: 9025e62758f5d07844a35073ca48f95068e5ae02