Polish target and client cancellation behavior
When we dispose targets and clients based on project name, we iterate through all relevant futures and attempt to dispose them. However, we didn't try to catch exceptions when calling "get" on the futures. This resulted in crash reports (see bug). With this CL, we dispose the clients that are done, and cancel those that are still in progress, and ignore those that failed.

I also took the liberty to refactor projectName out of AppInspectionTarget, because it doesn't need it.

Test: Updated AppInspectionTargetManagerTest and add new AppInspectionTargetTest.
Bug: 158931109
Change-Id: I895e9a1cc1c5c6a0e4a418342af3abf06dd27fea

GitOrigin-RevId: 19a2a91b1f721cba6502dd37b5c830c30d1268e2