Fix mouse hit precedence
When you hover or click on the device view, the click/hover should go to
the view whos owner is at the highest level, rather than simply the view
that's at the highest level. This means that if a parent draws above a
child, clicks will go to the child rather than the parent. This is
what's expected in the normal case, since usually parents don't
completely obscure children, but it is frequently the case that they do
some drawing on top.

Note that simply prioritizing based on non-draw hierarchical depth isn't
correct, since it's possible for a view to be drawn on top of one that
is, in terms of non-draw hierarchy, farther from the root. For example,
consider this view tree A -> [B -> [C], D], where all the views are
drawn on top of each other, in that order. A click should go to D, since
it is drawn on top, but would if using non-draw depth would go to C,
since it's deeper.

Test: added a test
Bug: none
Change-Id: Ic6bb0f21e9556156e7b8fe378e9459b6be62ea63