Bomb slow muted failing tests in IDEA to speed up test suite
This commit is to experiment with Android test suite performance. Most like it will not speed up anything. Gradle test cases are slow because of indexing. But in fact, sync tests do not need any indexes to complete. Mechanics is as follows: test case imports project from gradle. During import new modules are created. When a module is committed, it triggers "roots changed" event. This triggers indexing. Normally indexing is happening in parallel background thread, but there is a check in `com.intellij.openapi.project.DumbServiceImpl.isSynchronousTaskExecution` which detects unit-test mode, and starts execution in EDT. So test case have to wait few more minutes. Subsequent runs are much faster when indexes are warmed up.

GitOrigin-RevId: 0d183a3cc65b3e2583d440483e89f73e060af56c