[Gradle DSL] Introduce InterpolatedText.
In the DSL, we always treated interpolated text values as String and this was fine for groovy because of the flexibility of it's syntax. As KTS is not supported, and when setting an value to an interpolated String, we need to make sure of two things:
- The injection is resolvable (i.e. presents a reference to another property available from the context of the property concerned)
- The injection value is written in the correct syntax (which is not always obvious in KTS)

This change introduces the new API class InterpolatedText, and is the first step to enhancing support of interpolations at the DSL level. Following changes contain more details about the integration with other components.

Bug: 142454204
Bug: 148283067
Test: N/A

Change-Id: Ib5e8cb2bd1aff22fb17c378c8666c4aecc5afb31

GitOrigin-RevId: 982075ec2d5da5176a0c48f602d3e8692bbfcca7