[AUA] usages in KotlinScript Gradle build files
We currently display a usage type of "... in Gradle build script" for
Groovy usages in build files, but usage types corresponding to Kotlin
semantics in KotlinScript build files.

Fix the inconsistency by checking whether we are in a Gradle build
file for Kotlin elements too, and further place the
GradleUsageTypeProvider at the front of the list where it will take
priority over the Kotlin plugin's UsageTypeProvider.  (Ideally these
usage type providers would have ids in the plugin definitions so that
we could use order="before <id>", but they don't).

Bug: n/a
Test: existing, added
Change-Id: I6a0f1ec351c39483caba30f8cf1b70b3afcfd836

GitOrigin-RevId: 898e90c580e1bc23e68ed6e1ba14e00993a1da02