[AUA] customize UsageTypes for better preview
for each usage, provide a usage type that reflects the change that
will be applied to that usage specifically (as opposed to the command
in general).  For small refactoring components this may be an
additional layer of noise for not very much gain, but for refactorings
with non-trivial settings and multiple different areas of activity,
being able to group them by action will be helpful.

We strain at the edge of the customizeability of refactoring
processors and usage views here.  Fundamentally, there is an
assumption in the UsageView and related classes that a given
PsiElement is associated with at most one Usage: this allows
getUsageType() to operate even when passed only the PsiElement.

In our case, we will often have several Usages associated with a
single PsiElement: for example, when adding multiple properties to the
same block.  This is incompatible with customizing the usage types
directly, so we take the only solution open to us: we systematically
wrap the PsiElements associated with the usages in our own class,
which mostly acts as a delegator but implements and overrides enough
to function within the UsageView apparatus.

Then, subclass UsageTypeProvider and hook in appropriately, where
"appropriately" means "be careful to override the "Usage in Gradle
build file" usage type provider.

Bug: 149565531
Test: manual
Change-Id: I2b2ab7f28497a115281958eb6162160a2d73c2a0

GitOrigin-RevId: a74bbdcaeb5d029b1b8dff7869de72ead0a7358b