Remove obsolete abstract class for memory stage view's layout
This change introduces no behavior change. It simply removes
indirections and inlines private helpers.

The abstract class for memory stage view's layout used to be there as a
temporary solution for choosing how to lay out the stage, in particular
the heap dump view, depending on the feature flag. But this has been
obsolete since the heap dump was implemented in its own stage.

Bug: n/a
Test: existing
Change-Id: I0dde2c6dbf27c79baa4e80dae81d9363188f27ab

GitOrigin-RevId: 73a4194fab40ed2b90ac3b25924798bdbbfdd8d2