Use JDK from STUDIO_GRADLE_JDK in IdeSdks#getJdk
... when it is defined, valid and has not been changed in the PSD.

When the environment STUDIO_GRADLE_JDK variable is set to a valid jdk
location, the PSD uses its value to display as currently used JDK but
when we invoke Gradle, we are not using its value.

This cl modifies IdeSdks#getJdk to first check if the jdk pointed by the
environment value should be used.

Bug: 161405154
Test: IdeSdksTest
Change-Id: Id38f9cb0dc2211627abfe9240e1952c3a367e276

GitOrigin-RevId: 698c9cd1b30c8931f813c831f38f822647a55118