Update lint fix data map to be persistable
Most lint quickfixes are described as document edits or XML operations
as well as logical operations like reformat or add import.

However, there are some cases where the fix is more complicated and
lives in the IDE codebase, and the lint check simply passes a map of
keys+values to the IDE for interpretation. In some cases, this would
include data like PsiElements.

This CL reacts to the upstream lint API changes to the datamap API.

Test: Existing (and on the IDE side)
Bug: None
Change-Id: I1cca74dc5dc983dd4d4ba5327f5ec03581c397ae

GitOrigin-RevId: a4ee85a34d30ff30b20ad8d643d5968c6c236d0c