Optimisation for TagToClassMapper.isUpToDate()
When we are checking if a class map is up to date, on
every kotlin/java source change. We perform a check that
ends up using isInheritor. This is unnecessary in most cases,
as we already know if a class inherits from View, or
Preference, as those classes are the keys in the TagToClass
map.

This speeds up isUpToDate() on source changes from 120-200ms
to <2ms on empty projects. This also extends the previous
optimisation on computing the initial map to Preferences files.

Bug: 158358989
Test: existing
Change-Id: Id66a809ee8b887408e53d3bb1bd2cb77804c434e

GitOrigin-RevId: ec4fdca25421ae547ca0b44a42ad04e28c6d2b0b