Add loop playback action to the animation inspector
Add a loop toggle action to the animation inspector that, when selected,
keeps the timeline running indefinitely. When repeatable animations are
being tracked, we use the timeline as a sliding window, adding an offset
everytime the timeline reaches its end, until all the iterations are
covered. At that point, we return to the beginning. If all the tracked
animations fit the timeline (a.k.a. window size), we simply return to the
beginning when reaching the end.

Bug: 162015929
Test: Manual. UI test can be done once we merge dev16.
Change-Id: I97f097d75af65c3a524862610381f154c8dc715d