Create second action in ActionHandleTargetTest
The original issue was that a second action cannot be created
from an action handle that already has an outgoing action that
is selected. This issue was fixed by ag/12123063 which changed
actions from targets to hit providers.

Adding this case to the action handle target test anyway since
we weren't testing it. Also going through the interaction
manager instead of calling directly into the target to match
the real scenario more closely.

Bug: 160195388
Test: self
Change-Id: I963242f89848fe060da1b2d827cd59802937c540

GitOrigin-RevId: c0ff6c9999a6957d68436181786b32418b19bce4