[lint] visit slightly more expressions
When checking for the absence of something in order to signal a
problem, we need to be more certain that we have searched all the
different ways that that something could be expressed.  In the
particular instance being dealt with here, language level properties
in java-plugin modules could be expressed as

  java {
    sourceCompatibility = ...
  }

which is already handled by LintIdeGradleVisitor, but also as

  java.sourceCompatibility = ...

or indeed in Groovy

  java.sourceCompatibility ...

Extend LintIdeGradleVisitor to consider top-level application
statements and assignments where the expression is qualified, and
extract the qualifier as the parent and the name as the property name.

Bug: 161877341
Test: added in LintIdeGradleDetectorTest
Change-Id: Ia77ac37473c30d5fef7a158d8805a2cc2a818c2b

GitOrigin-RevId: 66c27ead460d8219c811a11e5554ddb3982ab30a