[AUA] rename the "Do Refactoring" button
Doing this requires overriding BaseRefactoringProcessor's
customizeUsageView method, which is the closest overridable method to
the presentation of the usages view.  It's not particularly close, and
there aren't hooks to allow running the parts of the base
functionality that we want, so we have to reimplement it for
ourselves.  This has the potential to lead to inconsistent experience
further down the line if the two codepaths diverge.

Bug: 162326171
Test: manual
Change-Id: Iffb1438ae68550c8f1bc1586c36bb9a4775d60f8