[AUA] rearrange Java8 dialog code
Separate out the initialization of the processor and related fields
from the UI and component initialization and display.

This allows an alternative flow through the dialog, where we will
return to it from an already-established processor (from the preview
window).  In that case, we should not re-set the enabled status of any
processors, and we should default the noLanguageLevelAction to
whatever the user specified previously.

(The return to the dialog will be actually implemented in a subsequent
change.)

Bug: 162326239
Test: added
Change-Id: Id7ca50771cebc1d9d03e8fd1a699d9c0c3ea666e

GitOrigin-RevId: c926924280efa98fe2b612171b888c4b8f521171