[Dsl,.gr] parsing of language level application statements
The sourceCompatibility and targetCompatibility application statements
are available in Groovy in the context of the java Dsl element /
configuration closure, but not in the apparent short-circuit of the
configuration closure at toplevel.

Both use cases are implemented using the JavaDslElement, which must
therefore have additional support to perform effective canonization of
application statements as GradleDslLiterals only in the context where
they are valid.

(Supporting parsing the full set of ways to set the language level is
necessary for the change to running JDK 8 by default.)

Bug: 162325061
Test: added
Change-Id: Ice8a453b4ac8b53fed1cd403b6d1a859a1347fe3

GitOrigin-RevId: 1eed0ff4168c4a11c9ea85f63b4032226ef73867