[AUA] add back button to preview
The usageView must remain open while in the dialog/preview cycle,
because the refactoring processor re-uses it, removing and adding
usages in bulk.  This means that in principle when we get to a point
when we wish to Upgrade from the dialog having seen a preview, we have
to close the usageView ourselves ("Cancel" and "Do
Refactoring"/"Perform Upgrade" having the internal machinery to do
this internally).

It is safe to close the usage view and dispose of its resources from
the Upgrade button, because that reruns findUsages() from scratch,
rather than using any usage information from the usageView.

Fixes: 162326239
Test: manual
Change-Id: Ia4e4aec4b970c6734e25e8fea3f8049733db16ff

GitOrigin-RevId: 1467f5ee8ff6d4f60e16e7117931975a81c4b5a8