Adds Stamper tool to stamp distribution files
This tool knows how to update xmls with build info.
Initially we use the same logic used by IJ's groovy
scripts to maintain parity, but they are fragile and
buggy (We ship the android/plugin.xml with version
"dev build" because it does not get stamped)

Once we transition, we can improve this mechanism.

Test: N/A
Bug: 156497102
Change-Id: I35d501c12ef21495d227b26f95b9ff3ff772e356

GitOrigin-RevId: 6f1b6b4def8c4cedb26a2c82f5b087c984c68013