Avoid showing 0GB for moderately small legends
The formatter is showing up to 1 digit after the decimal as a balance
between precision and conciseness. At the same time, it's also
displaying all memory legends in GB when the axis is 1GB+.
When several legends are less than 100MB, this can be confusing, as
they all show up as `0.0GB` even if their sum is significant.

This change allows customizing the max number of digits after decimal
based on the value. For the memory axis, we try to keep at least one
significant digit if the value is `0.001` or more.

Fixes: 160830139
Fixes: 160359105
Test: add new
Change-Id: I0e86bc2c846a4a7a6815b33ce99faeb94dc155cd

GitOrigin-RevId: 5686fe4ac324630ebc7cd4173c2a7cf6941d030c