Don't resolve aliases when parsing AndroidManifest.xml
Problem: In the context of building a list of candidate activities
to establish the "Default Activity", it is a mistake to resolve
aliases found int AndroidManifest.xml. If resolved, the
Intents of the alias are discard in favor of the target activity.

Solution: Treat "activity-alias" node exactly like "activity" node.

Fixes: 162259273
Test: Unit Tests
Change-Id: Ia15e8825da6c9db4986103fa26b188bc2613266e

GitOrigin-RevId: 7cd77350f8cb179826aee96b65e30089cd81a255