Reuse JARs that have only changed timestamps
This change uses CRCs to quickly check if the content
of a .class file has not changed. If this proves not
to be enough we can do a more expensive comparison.

But as is, this makes it easier to reuse artifacts and
see smaller diffs between updates.

Additionally only use the major IJ version to avoid
changing directories all the time

Test: N/A
Bug: N/A
Change-Id: Iad2bbcf5ee0808082b4d6e6a1e2a66fa9195a9fb