Fix IssuePanel causing UI freezes
The issue panel uses HTML to render error descriptions. Unfortunately,
the component used to render these, can be extremely slow to layout.
This CL avoids the component being populated when not visible so it
saves the cost in the most common sceneario.

Additionally, fixing an NPE in the issue panel caused by the
accesibility feature.

Test: Manually verified
Bug: 162809891
Fixes: 162400043
Change-Id: Idd7528c19609c3ebff8eb42060e0ce4f799583a5

GitOrigin-RevId: 3d87e387063c00a2609eabf33defa1824c959fa5