[Dsl] flavor collection known methods
Groovy's Dsl will automatically create contained objects if an
unrecognized method is called on a named domain object container.
This means that for accurate parsing, we need to know the set of known
methods.  Support for handling the special case already exists, but
users are tripping against some entries which are not in the known
method names set; add some more (and alphabetize so that later
maintainers, or future me, can see at a relative glance if something
is missing).

Fixes: 162351817
Test: added
Change-Id: Id8d2fff3359d84872f270c84004597e7eb1f2c71