[AUA] add explanatory comment by jvmTarget
Users might express surprise that when inserting directives to
continue using Java 7, kotlinOptions.jvmTarget is set to 1.6.  This is
correct, in that jvmTarget does not support 1.7, but leaving the user
to puzzle that out (probably by "correcting" the value to 1.7 and then
discovering that the project does not build) will lead to many
duplicate bug reports resolved as Won't Fix (Intended Behaviour).
To forestall that deluge of bug reports, add a comment by the
jvmTarget output to reassure the user that the value is deliberate.

This happens to be a point where we actually need the speculative
generality of the previously-unused SpoilingGradleBuildModelUsageInfo:
it is essentially impossible to modify comments through the build
model, or even (in a language-independent way) through the Psi.
Luckily, both of our currently-supported build language support line
comments preceded by `//`, so we insert a character sequence into a
document, and make sure to save and commit.

Fixes: 162516142
Test: existing tests adapted
Change-Id: Iae9ecabf2d389f74157892a7fc839918dce9f896

GitOrigin-RevId: beb284178243c421ce873da01f5aad1ea0e27eeb